Add files via upload
Improvements made to logic and readability of Menu class.

Added new functions to add a cursor in menus that responds to keyboard input to select menu options. It snaps to the same locations as the text on screen and is dynamically created so will change along with any added or removed menu options.

Added an Escape option to return to main menu from submenus.