Add files via upload
Finished converting all rendering of text into the text_setter and create_rect_with_pos functions.

Began overhaul of game's logic:
Created a Game Class which now creates menu and level objects.   
Have begun separating classes wherever possible. Menu now does not directly reference or alter any variables of other classes.
Removing all such references and alterations of other variables in Level is a work in progress. Have moved adjusting of the concurrent_bullets and cooldown_rate to the Player class.

Moved show_highscores() from a part of the MenuSystem class to the Game Class, which will show the highscores of the Endless game, but otherwise won't be created, and will directly create a new menu object. 

Old Menu's and Levels are destroyed upon the creation of each other - i.e. (menu created -> level deleted). 

The improved logic allowed the temporary kill_player function to be drastically reduced - over half of the reset variables are removed so far.

Gauntlet Mode is now created and somewhat working. It will load any level correctly but does not progress to the next level as of yet.